fix: MSVC build エラー修正一括
複数の MSVC エラー/警告を修正:

1. C1083 (削除済みソース参照):
   提案 5 で削除した timed-effect ファイル群を VS プロジェクトファイル
   (.vcxproj / .vcxproj.filters) から除去:
   - player-acceleration/blindness/confusion/deceleration/fear/
     hallucination/paralysis/poison/protection.cpp/.h
   - timed-effects.cpp/.h

2. C4458 (パラメータがクラスメンバをシャドウ):
   - creature-entity.cpp:558/569 get_inven_cnt/get_equip_cnt() の
     ローカル変数 count → cnt にリネーム (CreatureEntity::count をシャドウ)

3. C4242 (int→narrower 型への暗黙縮小変換):
   - CreatureEntity::get_level() を `int` から `PLAYER_LEVEL`
     (int16_t) 返却に変更。これで PLAYER_LEVEL 型代入の警告
     20+ 箇所を一括解消
   - score-util.cpp / inventory-curse.cpp / spells-trump.cpp:
     std::min<short>(get_level(), N) を std::min<int>(...) に
   - monster-processor.cpp: set_timed_effect 引数を short にキャスト

4. C4459 (ローカル変数がグローバルをシャドウ):
   - cmd-pet.cpp 内 target_pet → pet_monster にリネーム
     (グローバル target_pet input option をシャドウしていた)

5. C4805 (bool == flag_cause):
   - permanent-resistances.cpp の any_bits(has_levitation(), ...)
     を get_levitation_flags() 経由に変更
   - CreatureEntity に get_levitation_flags() virtual を新規追加

+++ src/inventory/inventory-curse.cpp
@@ -440,7 +440,7 @@ static void curse_drain_mp(CreatureEntity &creature)
     const auto *item_ptr = choose_cursed_obj_name(creature, CurseTraitType::DRAIN_MANA);
     const auto item_name = describe_flavor(creature, *item_ptr, (OD_OMIT_PREFIX | OD_NAME_ONLY));
     msg_format(_("%sはあなたの魔力を吸収した！", "Your %s drains mana from you!"), item_name.data());
-    creature.sub_csp(std::min<short>(creature.get_level(), 50));
+    creature.sub_csp(std::min<int>(creature.get_level(), 50));
     if (creature.get_csp() < 0) {
         creature.set_csp(0);
         creature.csp_frac = 0;
@@ -535,5 +535,5 @@ void execute_cursed_items_effect(CreatureEntity &creature)
         msg_print(_("なにかがあなたの体力を吸収した！", "Something drains life from you!"));
     }
 
-    take_hit(creature, DAMAGE_LOSELIFE, std::min<short>(creature.get_level(), 50), _("審判の宝石", "the Jewel of Judgement"));
+    take_hit(creature, DAMAGE_LOSELIFE, std::min<int>(creature.get_level(), 50), _("審判の宝石", "the Jewel of Judgement"));
 }
