+++ src/spell-realm/spells-trump.cpp
@@ -191,7 +191,7 @@ void cast_shuffle(CreatureEntity &creature)
         msg_print(_("《恋人》だ。", "It's the Lovers."));
 
         if (const auto dir = get_aim_dir(creature)) {
-            charm_monster(creature, dir, std::min<short>(creature.get_level(), 20));
+            charm_monster(creature, dir, std::min<int>(creature.get_level(), 20));
         }
 
         return;
